+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 
 - Update pystac from 0.5.6 to 1.1.0
 - Move stac-validator 2.3.0 into repository
+- Best practices check for too many links in object
 - Best practices check for summaries in collections
 - Validation from stac-validator 2.3.0  
 - Links and assets validation checks