+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v0.1.0] - 2021-11-26
+## [v0.1.0] - 2021-11-26 - 2021-12-05
 ### Added
 
+- Best practices ensure item ids don't contain ':' or '/' characters  
+  https://github.com/radiantearth/stac-spec/blob/master/best-practices.md#item-ids
 - Best practices check for item ids to see if they match file names
 - Add url support, check for valid urls, validate urls
 - Add pystac validate_all to new cli option -> recursive